Move socket.io event stuff to parent of parent and share loglines between children instead of having multiple connections from dryrunprogressinfo component
