+++ public/components/ConfigChange/ConfigChangeStep4.js
@@ -56,6 +56,7 @@ class ConfigChangeStep4 extends React.Component {
             dryRunJobStatus={dryRunJobStatus}
             dryRunProgressData={dryRunProgressData}
             jobId={jobId}
+            logLines={this.props.logLines}
           />
         </div>
       </div>
